OOP Implemented
Created a Game class that runs a game/round for the user. This class will be used to archive past games and act as a record for previous games played in the future. This included me adding a timer to see how long it took the user to complete the game, a counter of incorrect answers the user had, and new outputs at the end of the game. Game still console locked and only runs once. Made it to where game wont crash if user doesn't answer an equation with a valid integer.